backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
fakeredis
pandas
numpy
faiss-cpu